requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
langchain-google-genai
python-dotenv
tabulate